requirements.txt
PyInstaller>=3.2
PyQt5>=5.8
gitpython==2.1.3